Emit .html files in the write phase
A few tricks had to be pulled off to make this possible, and this is by
no means a presentable product yet. The way it's done is that we set up
a fake DOM context using jsdom, then load in the bundles compiled by
webpack for the UI core and every UI plugin that is used. Then, we
launch the UI into a "static" mode where it exposes an API for rendering
a page at some URL to a _string_ using the React.renderToString() API.
That DOM string is then serialized to a file, along with the outer HTML
(index.tmpl.html) needed for the page.

The remaining issues have to do with the URLs and hash-tags. This is
definitely a WIP still.