Generate URIs for corpus documents
References #47

HTMLSerializer will now visit all the nodes in the corpus and generate a
URI for each document. A namespace (or document) can specify/override
the default URI if it needs to (e.g. useful for plugins that nest their
resources under certain fixed routes, like "/modules/" for JS).

The generated @href meta attribute does _not_ contain the file
extension. Instead, we append/remove the extension at the end-points as
needed; the UI and the file emitter.