Ignore baseURL in rendered markdown links in MPM
The anchors in multi-page mode do not need to be fully qualified; we
ignore the baseURL. The baseURL passed to the renderer must be the
fully-qualified path (i.e. the href of the node).

So, for a document with the href "/markdown/readme" and a heading that
reads "Usage Examples", we will have the following anchors:

- SPM: /index.html#/markdown/readme/usage-examples
- MPM: /markdown/readme.html#usage-examples