Revisit URL encoding and allow other rewriting schemes
Rewriting may now be done using UID, URL, and filepath.

We no longer encode legal symbols in URLs like "#". This is done with
the use of encodeURI() instead of encodeURIComponent().

This might raise issues if the documents have illegal filename
characters in their names, but so far this has been restricted to
entity names which are not serialized to file anyhow.