Symlink dist under packages/tinydoc/dist
Node v6 started failing on travis because it wouldn't find the files
under dist/ which are needed by the integration tests. I'm guessing it
has something to do with `npm link` internals.

The patch also disables node v0.12 on travis because we don't have jsdom
on that version. Perhaps we should downgrade to jsdom 6.x for 0.12
support or just drop 0.12 support altogether. A topic for the future.