[html] improve 404 experience & compile inline scripts
It's once again possible to inject custom inline scripts. Right now
they're bounded to a "notFoundComponent" option which will generate an
outlet occupant with the specified component.

The 404 screen will now display a link that will take the user back
to the closest working document or /index.html if none was found.