drop reduce routines, add summaryFields
"reduceRoutines" that were being passed to each reduceFn have been
removed. They were there in order to inject reducers with the ability to
render markdown and stuff, but now we've delegated that to the
TreeRenderer where it will inspect every node for the "summaryFields"
property and extract the summary from them if possible.