apply tree merging right after parsing
fixes GH-113

This patch takes a different approach to merging change trees; instead
of having to fiddle with render and tree operations (and reduced
documents, for the matter), we instead apply the merge right after
parsing and right before refinement.

This design has the benefit of not exposing this detail to consumers
(e.g. processors) as they will still have the complete list of documents
to process in all stages (whether it's an incremental or full run is
transparent to them now.)